+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 ## Next - May XXth
 - button to quickly scroll to the beginning and end of collection screen
 - add name filter indicator in header
+- setting to reset name filter when pegasus reopens
 
 ## Next - May 23rd
 - new settings filters